remove redundant test setup code
